Fixed KeyError: 'prediction'
Fixed KeyError: 'prediction' by giving the validation_data a new column with the predictions. I also updated google colab because it wasn't working, and I've updating it with the basis of using my code. I've noticed that the MMC numbers are incorrect, so if you approve the code, please fix that.